+++ configuronic/__init__.py
@@ -1,4 +1,4 @@
 from .cli import cli, get_required_args
-from .config import Config, ConfigError, config
+from .config import Config, ConfigError, ImportNotAllowedError, config
 
-__all__ = ['config', 'Config', 'ConfigError', 'cli', 'get_required_args']
+__all__ = ['config', 'Config', 'ConfigError', 'ImportNotAllowedError', 'cli', 'get_required_args']